requirements.txt
fastapi
uvicorn
pypdf
langchain-text-splitters
cohere
python-dotenv
numpy
python-multipart
